Prepare for proper nested transaction handling.
Move all the nested level checking into the txn_{begin,commit,rollback}
functions, which then decide to call _dbh_txn_{begin,commit,rollback}
or new functions _dbh_txn_nested_{begin,commit,rollback}, depending
upon what commit level they are operating at.  By default the
functions _dbh_txn_nested_{begin,commit} do nothing and are just
stubs, and the nested rollback causes an exception.

The _dbh_txn_nested_* functions are candidates for over-riding in
the .../DBI/database.pm files so that each driver can implement
nested transactions if they can.  For example the mysql driver can
support a level of nested commits using the SQL syntax,

    SAVEPOINT identifier
    ROLLBACK [WORK] TO SAVEPOINT identifier
    RELEASE SAVEPOINT identifier,

and so with these changes to .../Storage/DBI.pm it is a relatively
simple matter to implement this.

Secondly remove the dependancy upon the DBI AutoCommit variable,
so that the transaction code trusts DBI's judgement as to whether
a transaction command is valid or not, and propagates any errors
back to the caller.